Refactoring in RTP reader code: sequence number wrapping logic moved to RtpReader
BUG=None
TEST=Unittests

Review URL: http://codereview.chromium.org/5110008

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@67212 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 8e88605290262afe6291f997abc0d5ff9197c595

+++ remoting/protocol/rtp_video_reader.h
@@ -52,7 +52,7 @@ class RtpVideoReader : public VideoReader {
   RtpReader rtp_reader_;
 
   PacketsQueue packets_queue_;
-  uint16 last_sequence_number_;
+  uint32 last_sequence_number_;
 
   // The stub that processes all received packets.
   VideoStub* video_stub_;
